created two components poorly
