+++ utils/table.css
@@ -1,0 +1,35 @@
+h1 {
+  padding-top: 15px;
+  padding-bottom: 15px;
+}
+td {
+  padding-bottom: 20px;
+}
+h3 {
+  padding-top: 20px;
+  width: 160px;
+  color: #151b54;
+}
+h5 {
+  text-decoration: underline;
+  color: #151b54;
+}
+button {
+  color: black;
+  background-color: lightgray;
+  font-size: 20px;
+  text-align: right;
+}
+td {
+  vertical-align: top;
+  height: 150px;
+  border: 2px solid #151b54;
+  text-align: center;
+}
+th {
+  text-align: center;
+}
+h6 {
+  font-size: 20px;
+  color: #4863a0;
+}
